+++ CHANGES.md
@@ -11,6 +11,11 @@
 ## Develop
 
 * World trade codes are now recalculated when a UWP value is changed.
+* Profession specialisation skills are now displayed if they are at level zero but trained.
+* Skill training macro now supports adding new skills and specialisations.
+* CSS for journal headings now focused on the page content, rather than the entire journal window
+  (it was overriding the window title bar as well, which was undesirable).
+
 
 ## 0.13.0 (Beta)